Move OrderBy push-down optimization for joins.
Until now the OrderBy push-down optimization took place in the Analysis phase,
and if a successful pushdown was applied it prevented the re-ordering of the
relations of the join in the LogicalPlanner which in turn could prevent the
usage of new join algorithms (eg. hash-join).

This changes moves the OrderBy optimization from the Analysis to the
LogicalPlanner. So now the OrderBy optimization is done through the
`tryOptimize()` mechanism which is already in place.

The OrderBy push-down can only be applied if the symbols of the OrderBy refer
to the top-most left relation of the join tree. This condition is still valid
after this change but the join tree is allowed to be already re-order in an
optimal way based on the join conditions (apply as many join conditions as
early as possible). In the near future the join tree can already be created in
an optimal way also because of the join algorithm chosen and on top of that we
can still try to apply the OrderBy push-down.