Optimize HashJoin execution by switching sides of relations.
After benchmarking we concluded that the smaller table should be on the right side
of the hash join. In case the left table is loaded into memory in blocks then the right side
is repeated, therefore is better to be the smaller one. In case the left table fits into one block
and there is no repetition of the right one then the benchmarks showed that performance is slightly
better if smaller table is on the left side, but the difference is minimal.

So as a conclusion we switch tables when building the execution
plan to ensure that the smaller table is on the right side.